add tests #2
