+++ kafka-connect-http-connectors/src/test/java/io/github/clescot/kafka/connect/http/HttpTaskTest.java
@@ -1,6 +1,7 @@
 package io.github.clescot.kafka.connect.http;
 
 
+import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
@@ -20,6 +21,9 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.runner.RunWith;
 import org.testcontainers.shaded.com.google.common.collect.Maps;
 
+import java.io.IOException;
+import java.net.DatagramSocket;
+import java.net.InetAddress;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
@@ -28,9 +32,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(Enclosed.class)
 public class HttpTaskTest {
-
-    private static final HttpRequest.Method DUMMY_METHOD = HttpRequest.Method.POST;
-    private static final String DUMMY_BODY = "stuff";
+    public static final String OK = "OK";
     private static final String DUMMY_BODY_TYPE = "STRING";
     @RegisterExtension
     static WireMockExtension wmHttp = WireMockExtension.newInstance()
@@ -47,6 +49,15 @@ public class HttpTaskTest {
         HttpTask.removeCompositeMeterRegistry();
     }
 
+    private String getIP() {
+        try (DatagramSocket datagramSocket = new DatagramSocket()) {
+            datagramSocket.connect(InetAddress.getByName("8.8.8.8"), 12345);
+            return datagramSocket.getLocalAddress().getHostAddress();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     @Nested
     class Call {
 
@@ -54,13 +65,24 @@ public class HttpTaskTest {
         void test_nominal_case() throws ExecutionException, InterruptedException {
             //given
             WireMockRuntimeInfo wmRuntimeInfo = wmHttp.getRuntimeInfo();
-
+            WireMock wireMock = wmRuntimeInfo.getWireMock();
+            String bodyResponse = "{\"result\":\"pong\"}";
+            wireMock
+                    .register(WireMock.post("/ping")
+                            .willReturn(WireMock.aResponse()
+                                    .withHeader("Content-Type", "application/json")
+                                    .withBody(bodyResponse)
+                                    .withStatus(200)
+                                    .withStatusMessage(OK)
+                                    .withFixedDelay(1000)
+                            )
+                    );
             Map<String,String> settings = Maps.newHashMap();
             HttpConfigDefinition httpConfigDefinition = new HttpConfigDefinition(settings);
             HttpConnectorConfig httpConnectorConfig = new HttpConnectorConfig(httpConfigDefinition.config(), settings);
             HttpTask httpTask = new HttpTask(httpConnectorConfig,new OkHttpClientFactory());
 
-            HttpRequest httpRequest =  getDummyHttpRequest("http://127.0.0.1:"+wmRuntimeInfo.getHttpPort()+"/path2");
+            HttpRequest httpRequest =  getDummyHttpRequest("http://"+getIP()+":"+wmRuntimeInfo.getHttpPort()+"/ping");
             HttpExchange httpExchange = (HttpExchange) httpTask.call(httpRequest).get();
             assertThat(httpExchange).isNotNull();
             assertThat(httpExchange.getHttpRequest()).isNotNull();
@@ -98,7 +120,7 @@ public class HttpTaskTest {
 
     }
     static HttpRequest getDummyHttpRequest(String url) {
-        HttpRequest httpRequest = new HttpRequest(url, DUMMY_METHOD);
+        HttpRequest httpRequest = new HttpRequest(url, HttpRequest.Method.POST);
         Map<String, List<String>> headers = com.google.common.collect.Maps.newHashMap();
         headers.put("Content-Type", Lists.newArrayList("application/json"));
         httpRequest.setHeaders(headers);
@@ -107,14 +129,4 @@ public class HttpTaskTest {
         return httpRequest;
     }
 
-
-    private String getLocalHttpRequestAsStringWithPath(int port, String path, String method, String dummyBody) {
-        return "{\n" +
-                "  \"url\": \"" + "http://localhost:" + port + path + "\",\n" +
-                "  \"headers\": {},\n" +
-                "  \"method\": \"" + method + "\",\n" +
-                "  \"bodyAsString\": \"" + dummyBody + "\",\n" +
-                "  \"bodyType\": \"" + DUMMY_BODY_TYPE + "\"\n" +
-                "}";
-    }
 }
